requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
alembic>=1.12.0
pre-commit>=3.5.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1  # Required for FastAPI TestClient
pytest-cov>=4.1.0  # For test coverage reporting

# Code formatting dependencies
ruff>=0.1.0
